Updating html template
Using bootstrap to create the template for the page
Added navbar,
main div to hold the blog posts including a header and footer for each
post. this will hold the title of the post, the auther, and timestamp
added side bar to list the last 5 posts and direct link to each one.