added angularModalService, updated app
added angularModalService to bower_components
this makes it convenient for using bootstrap modals with angularjs.
highlighted posts section was causing more problems through development,
so removed it and focused on completing other features so the app can
meet the requirements by the deadline.
used angular's form validation for the new post form.